Split S3TransferManagerAutoConfigurationTests from the main S3AutoConfigurationTests
