+++ spring-cloud-aws-autoconfigure/src/test/java/io/awspring/cloud/autoconfigure/s3/S3AutoConfigurationTests.java
@@ -30,22 +30,18 @@ import io.awspring.cloud.s3.S3ObjectConverter;
 import io.awspring.cloud.s3.S3OutputStream;
 import io.awspring.cloud.s3.S3OutputStreamProvider;
 import io.awspring.cloud.s3.S3Template;
-import io.awspring.cloud.s3.TransferManagerS3OutputStreamProvider;
 import io.awspring.cloud.s3.crossregion.CrossRegionS3Client;
 import java.io.IOException;
 import java.net.URI;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.test.context.FilteredClassLoader;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.lang.NonNull;
 import org.springframework.lang.Nullable;
 import org.springframework.test.util.ReflectionTestUtils;
-import software.amazon.awssdk.services.s3.S3AsyncClient;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.S3ClientBuilder;
 import software.amazon.awssdk.transfer.s3.S3TransferManager;
@@ -58,10 +54,10 @@ import software.amazon.awssdk.transfer.s3.S3TransferManager;
 class S3AutoConfigurationTests {
 
 	private final ApplicationContextRunner contextRunner = new ApplicationContextRunner()
+			.withClassLoader(new FilteredClassLoader(S3TransferManager.class))
 			.withPropertyValues("spring.cloud.aws.region.static:eu-west-1")
 			.withConfiguration(AutoConfigurations.of(AwsAutoConfiguration.class, RegionProviderAutoConfiguration.class,
-					CredentialsProviderAutoConfiguration.class, S3TransferManagerAutoConfiguration.class,
-					S3AutoConfiguration.class));
+					CredentialsProviderAutoConfiguration.class, S3AutoConfiguration.class));
 
 	@Test
 	void createsS3ClientBean() {
@@ -87,11 +83,13 @@ class S3AutoConfigurationTests {
 
 	@Test
 	void autoconfigurationIsNotTriggeredWhenS3ModuleIsNotOnClasspath() {
-		this.contextRunner.withClassLoader(new FilteredClassLoader(S3OutputStreamProvider.class)).run(context -> {
-			assertThat(context).doesNotHaveBean(S3Client.class);
-			assertThat(context).doesNotHaveBean(S3ClientBuilder.class);
-			assertThat(context).doesNotHaveBean(S3Properties.class);
-		});
+		this.contextRunner
+				.withClassLoader(new FilteredClassLoader(S3TransferManager.class, S3OutputStreamProvider.class))
+				.run(context -> {
+					assertThat(context).doesNotHaveBean(S3Client.class);
+					assertThat(context).doesNotHaveBean(S3ClientBuilder.class);
+					assertThat(context).doesNotHaveBean(S3Properties.class);
+				});
 	}
 
 	@Nested
@@ -112,10 +110,11 @@ class S3AutoConfigurationTests {
 
 		@Test
 		void createsStandardClientWhenCrossRegionModuleIsNotInClasspath() {
-			contextRunner.withClassLoader(new FilteredClassLoader(CrossRegionS3Client.class)).run(context -> {
-				assertThat(context).doesNotHaveBean(CrossRegionS3Client.class);
-				assertThat(context).hasSingleBean(S3Client.class);
-			});
+			contextRunner.withClassLoader(new FilteredClassLoader(S3TransferManager.class, CrossRegionS3Client.class))
+					.run(context -> {
+						assertThat(context).doesNotHaveBean(CrossRegionS3Client.class);
+						assertThat(context).hasSingleBean(S3Client.class);
+					});
 		}
 	}
 
@@ -123,15 +122,13 @@ class S3AutoConfigurationTests {
 	class OutputStreamProviderTests {
 
 		@Test
-		void whenS3TransferManagerInClassPathCreatesTransferManagerSS3OutputStreamProvider() {
-			contextRunner
-					.run(context -> assertThat(context).hasSingleBean(TransferManagerS3OutputStreamProvider.class));
+		void whenS3TransferManagerInClassPathDoesntCreateBean() {
+			contextRunner.run(context -> assertThat(context).doesNotHaveBean(S3TransferManager.class));
 		}
 
 		@Test
 		void whenS3TransferManagerNotInClassPathCreatesDiskBufferingS3OutputStreamProvider() {
-			contextRunner.withClassLoader(new FilteredClassLoader(S3TransferManager.class))
-					.run(context -> assertThat(context).hasSingleBean(DiskBufferingS3OutputStreamProvider.class));
+			contextRunner.run(context -> assertThat(context).hasSingleBean(DiskBufferingS3OutputStreamProvider.class));
 		}
 
 		@Test
@@ -188,10 +185,11 @@ class S3AutoConfigurationTests {
 
 		@Test
 		void withoutJacksonOnClasspathDoesNotConfigureObjectConverter() {
-			contextRunner.withClassLoader(new FilteredClassLoader(ObjectMapper.class)).run(context -> {
-				assertThat(context).doesNotHaveBean(S3ObjectConverter.class);
-				assertThat(context).doesNotHaveBean(S3Template.class);
-			});
+			contextRunner.withClassLoader(new FilteredClassLoader(S3TransferManager.class, ObjectMapper.class))
+					.run(context -> {
+						assertThat(context).doesNotHaveBean(S3ObjectConverter.class);
+						assertThat(context).doesNotHaveBean(S3Template.class);
+					});
 		}
 
 		@Test
@@ -218,63 +216,6 @@ class S3AutoConfigurationTests {
 						S3ObjectConverter converter = (S3ObjectConverter) ReflectionTestUtils.getField(s3Template,
 								"s3ObjectConverter");
 						assertThat(converter).isEqualTo(customS3ObjectConverter);
-					});
-		}
-	}
-
-	@Nested
-	class TransferManagerTests {
-		@Test
-		void createsS3TransferManagerBeanWhenInClassPath() {
-			contextRunner.run(context -> assertThat(context).hasSingleBean(S3TransferManager.class));
-		}
-
-		@Test
-		void usesExistingS3TransferManagerBeanWhenExists() {
-			S3TransferManager customDefinedS3TransferManager = Mockito.mock(S3TransferManager.class);
-			contextRunner.withBean("s3transferManager", S3TransferManager.class, () -> customDefinedS3TransferManager)
-					.run(context -> assertThat(context.getBean(S3TransferManager.class))
-							.isEqualTo(customDefinedS3TransferManager));
-		}
-
-		@Test
-		void doesNotCreateS3TransferManagerBeanWhenNotInClassPath() {
-			contextRunner.withClassLoader(new FilteredClassLoader(S3TransferManager.class)).run(context -> {
-				assertThat(context).doesNotHaveBean(S3TransferManager.class);
-			});
-		}
-	}
-
-	@Nested
-	class TransferManagerEndpointConfigurationTests {
-		@Test
-		void withCustomEndpoint() {
-			contextRunner.withPropertyValues("spring.cloud.aws.s3.endpoint:http://localhost:8090").run(context -> {
-				S3TransferManager transferManager = context.getBean(S3TransferManager.class);
-				ConfiguredAwsClient client = new ConfiguredAwsClient(resolveS3Client(transferManager));
-				assertThat(client.getEndpoint()).isEqualTo(URI.create("http://localhost:8090"));
-				assertThat(client.isEndpointOverridden()).isTrue();
-			});
-		}
-
-		@Test
-		void withCustomGlobalEndpoint() {
-			contextRunner.withPropertyValues("spring.cloud.aws.endpoint:http://localhost:8090").run(context -> {
-				S3TransferManager transferManager = context.getBean(S3TransferManager.class);
-				ConfiguredAwsClient client = new ConfiguredAwsClient(resolveS3Client(transferManager));
-				assertThat(client.getEndpoint()).isEqualTo(URI.create("http://localhost:8090"));
-				assertThat(client.isEndpointOverridden()).isTrue();
-			});
-		}
-
-		@Test
-		void withCustomGlobalEndpointAndS3Endpoint() {
-			contextRunner.withPropertyValues("spring.cloud.aws.endpoint:http://localhost:8090",
-					"spring.cloud.aws.s3.endpoint:http://localhost:9999").run(context -> {
-						S3TransferManager transferManager = context.getBean(S3TransferManager.class);
-						ConfiguredAwsClient client = new ConfiguredAwsClient(resolveS3Client(transferManager));
-						assertThat(client.getEndpoint()).isEqualTo(URI.create("http://localhost:9999"));
-						assertThat(client.isEndpointOverridden()).isTrue();
 					});
 		}
 	}
@@ -322,12 +263,6 @@ class S3AutoConfigurationTests {
 		public S3OutputStream create(String bucket, String key, @Nullable ObjectMetadata metadata) throws IOException {
 			return null;
 		}
-	}
-
-	@NonNull
-	private static S3AsyncClient resolveS3Client(S3TransferManager builder) {
-		return (S3AsyncClient) ReflectionTestUtils.getField(ReflectionTestUtils.getField(builder, "s3CrtAsyncClient"),
-				"s3AsyncClient");
 	}
 
 }
